These files are to train a monai UNet model
These three files need to be put together. Train_monai_Unet.py is used to train the model. The MONAI transform can be varied based on which variant to be used. The create_dataset_monai_unet.py is used to create patient and cancer-type stratified splits (train, validation, and test set) and create a data dictionary for MONAI training. The Inference_monai_Unet.py is for running inference using the trained model.